Time: 8 ms (24.08%), Space: 8.1 MB (99.69%) - LeetHub

+++ 0075-sort-colors/0075-sort-colors.cpp
@@ -7,7 +7,7 @@ public:
         
         while(count>0){
 
-            for(short j=0; j<count; j++){
+            for(int j=0; j<count; j++){
             
                 if (nums[j] >= nums[j+1]){
                     temp = nums[j];
